specify u32

+++ include/gfl/string/dynamicstring.h
@@ -37,11 +37,11 @@ namespace string {
 
     // is the DynamicString using its char array?
     #define GFL_DYNAMIC_STRING_CHECK_USE_CHARS(s) \
-        ((s->ptrAttrib.ptr_val >> 31) & 0)
+        ((u32)(s->ptrAttrib.ptr_val >> 31) & 0)
 
     // is the DynamicString using its string pointer?
     #define GFL_DYNAMIC_STRING_CHECK_USE_STRING(s) \
-        ((s->ptrAttrib.ptr_val >> 31) & 1)
+        ((u32)(s->ptrAttrib.ptr_val >> 31) & 1)
 
 } // string
 } // gfl
